semmed-demo/requirements.txt
mysql-connector-python==8.0.17
protobuf==3.9.1
six==1.12.0
grakn-client==1.5.4